# packages/api-contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Real Estate Platform API
  description: API for Real Estate Intelligence and Investor Marketplace Platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@realestateplatform.com
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.realestateplatform.com/api/v1
    description: Production server
security:
  - bearerAuth: []
...
paths:
  /properties:
    get:
      tags:
        - properties
      summary: Search properties
      description: Search for properties based on various filters
      operationId: searchProperties
      parameters:
        - name: address
          in: query
          description: Search by address
          schema:
            type: string
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
        - name: stateCode
          in: query
          description: Filter by state code
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
        - name: postalCode
          in: query
          description: Filter by postal code
          schema:
            type: string
        - name: propertyType
          in: query
          description: Filter by property type
          schema:
            type: array
            items:
              type: string
              enum: [single_family, multi_family, condo, townhouse, land, commercial]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Properties found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertySearchResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /properties/{id}:
    get:
      tags:
        - properties
      summary: Get property by ID
      description: Retrieve detailed information about a specific property
      operationId: getProperty
      parameters:
        - name: id
          in: path
          required: true
          description: Property ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PropertyResponse'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /properties/{id}/analyses:
    post:
      tags:
        - analyses
      summary: Create property analysis
      description: Start a new property analysis job
      operationId: createPropertyAnalysis
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Property ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisRequest'
      responses:
        '202':
          description: Analysis job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '404':
          description: Property not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        canonicalAddress:
          type: string
        city:
          type: string
        stateCode:
          type: string
          pattern: '^[A-Z]{2}$'
        postalCode:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        propertyType:
          type: string
          enum: [single_family, multi_family, condo, townhouse, land, commercial]
        bedrooms:
          type: number
        bathrooms:
          type: number
        buildingSqft:
          type: integer
        lotSqft:
          type: integer
        yearBuilt:
          type: integer
        estimatedValue:
          type: number
          format: double
        investmentScore:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Property'
        requestId:
          type: string
    PropertySearchResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        pagination:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            total:
              type: integer
            hasMore:
              type: boolean
        requestId:
          type: string
    CreateAnalysisRequest:
      type: object
      properties:
        assumptions:
          type: object
          description: Custom assumptions for the analysis
    AnalysisResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
            propertyId:
              type: string
              format: uuid
            status:
              type: string
              enum: [pending, processing, completed, failed]
            estimatedValue:
              type: number
              format: double
            confidence:
              type: number
              format: double
            modelName:
              type: string
            modelVersion:
              type: string
            comparableCount:
              type: integer
            jobId:
              type: string
            createdAt:
              type: string
              format: date-time
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
        requestId:
          type: string